requirements.txt
fastapi==0.111.0
alembic==1.13.1
pydantic[email]== 2.7.3
SQLAlchemy==2.0.4
SQLAlchemy-Utils==0.40.0
asyncpg==0.29.0
uvicorn==0.30.1
python-dotenv==1.0.1
pydantic-settings==2.3.1
PyJWT==2.8.0
cryptography==43.0.0
argon2-cffi==23.1.0
psycopg2-binary==2.9.10
celery[beat]==5.5.2
flower==2.0.1
redis==5.2.1